Add label configuration for main repos
This adds the label configuration for the repos `unit-e-desktop`, `docs.unit-e.io`,
`unit-e-clonemachine`, and `unit-e-project`.

Adding them in once place makes it easier to do changes across repos. The
`unit-e` label configuration maybe should also go here, but as that repo
is by far the most active one, it might be better to keep everything there
in one place, at least for now.

There is some duplication in the `labels.yaml` file which could be reduced
by more heavy use of YAML references and anchors. The file works as it is,
though, so this is left for later optimisations.

Signed-off-by: Cornelius Schumacher <cornelius@thirdhash.com>